requirements.txt
streamlit
openai
PyMuPDF
pymupdf4llm
python-dotenv
numpy
chromadb
PyYAML
trafilatura
html2text
requests
crawl4ai
llama-index-core
llama-index-llms-azure-openai
